+++ lib/core/src/display/shape/primitive/def.rs
@@ -1,0 +1,8 @@
+//! This module is the root module for all primitive shapes and shape transform definitions.
+
+pub mod sdf;
+pub mod class;
+pub mod transform;
+
+pub use sdf::immutable::*;
+pub use transform::immutable::*;
