+++ lib/core/src/display/shape/primitive/shader/helpers.glsl
